Add support for connecting to external services
This patch adds support for external services, that is services that
are not running in a Docker container.

If a service is in "EXTERNAL" mode, it will not be added to the
docker-compose file, but other services will try to connect to
that service by using an externalConnection in the connectionConfig.

Also, this patch reformats all code according to the default Kotlin
style.